pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "snow-town"
version = "0.1.0"
description = "Closed-loop learning ecosystem connecting Ultra Magnus, Sky-Lynx, and Agent Persona Academy"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.1.0",
    "mypy>=1.8",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["contracts*", "scripts*", "api*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true
